requirements.txt
pandas
sqlalchemy
psycopg2-binary
tqdm
requests
python-dotenv
ipython